Temp/Gemini-api-chatbot/Backend/requirements.txt
langchain==0.2.1
langchain-community==0.0.30
langchain-google-genai==1.0.1
google-generativeai==0.5.4
python-dotenv==1.0.1
pypdf==3.17.4
faiss-cpu==1.8.0
transformers==4.41.1
